Changes from Bullet.

+++ BulletSharpPInvoke/Dynamics/DynamicsWorld.cs
@@ -81,10 +81,10 @@ namespace BulletSharp
 
         public void AddRigidBody(RigidBody body, CollisionFilterGroups group, CollisionFilterGroups mask)
 		{
-            _collisionObjectArray.Add(body, (short)group, (short)mask);
+            _collisionObjectArray.Add(body, (int)group, (int)mask);
 		}
 
-        public void AddRigidBody(RigidBody body, short group, short mask)
+        public void AddRigidBody(RigidBody body, int group, int mask)
         {
             _collisionObjectArray.Add(body, group, mask);
         }
