+++ libbulletc/src/btDynamicsWorld_wrap.h
@@ -6,7 +6,7 @@ extern "C" {
 	EXPORT void btDynamicsWorld_addAction(btDynamicsWorld* obj, btActionInterface* action);
 	EXPORT void btDynamicsWorld_addConstraint(btDynamicsWorld* obj, btTypedConstraint* constraint, bool disableCollisionsBetweenLinkedBodies);
 	EXPORT void btDynamicsWorld_addRigidBody(btDynamicsWorld* obj, btRigidBody* body);
-	EXPORT void btDynamicsWorld_addRigidBody2(btDynamicsWorld* obj, btRigidBody* body, short group, short mask);
+	EXPORT void btDynamicsWorld_addRigidBody2(btDynamicsWorld* obj, btRigidBody* body, int group, int mask);
 	EXPORT void btDynamicsWorld_clearForces(btDynamicsWorld* obj);
 	EXPORT btTypedConstraint* btDynamicsWorld_getConstraint(btDynamicsWorld* obj, int index);
 	EXPORT btConstraintSolver* btDynamicsWorld_getConstraintSolver(btDynamicsWorld* obj);
